refactor(docx): extract revision review primitives
